+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed play button not visible anymore if buffering occurs while the player is paused
 - Fixed missing ':' between hours and minutes in video player
 - Fixed issue where the video is restarted when returning from another application
+- Don't hide player UI automatically if the video is buffering
 
 ## [0.9.0] - 2021-09-20
 ### Added